requirements.txt
streamlit==1.27.0
pymupdf==1.22.5
openai>=1.0.0
psycopg2-binary==2.9.7
pandas==2.0.3
python-dotenv==1.0.0